feat(invoice): stripe mock
