+++ openmeter/app/stripe/client/stripe_client.go
@@ -35,6 +35,11 @@ type StripeClient interface {
 	CreateCustomer(ctx context.Context, input CreateStripeCustomerInput) (StripeCustomer, error)
 	CreateCheckoutSession(ctx context.Context, input CreateCheckoutSessionInput) (StripeCheckoutSession, error)
 	GetPaymentMethod(ctx context.Context, stripePaymentMethodID string) (StripePaymentMethod, error)
+	// Invoice
+	CreateInvoice(ctx context.Context, input CreateInvoiceInput) (*stripe.Invoice, error)
+	UpdateInvoice(ctx context.Context, input UpdateInvoiceInput) (*stripe.Invoice, error)
+	DeleteInvoice(ctx context.Context, input DeleteInvoiceInput) error
+
 	GetClient() *client.API
 }
